% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{case_plot}
\alias{case_plot}
\title{Case plot for figures}
\usage{
case_plot(
  df,
  testing = FALSE,
  facet = "wrap",
  nrow = NULL,
  grid = FALSE,
  gridvar = NULL
)
}
\value{

}
\description{
Case plot for figures
}
\examples{

}
\author{
Lewis Spurgin
}
