% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist2_func}
\alias{dist2_func}
\title{Simulate social distancing using data within days}
\usage{
dist2_func(am, am.contacts, returns, dist.prop)
}
\arguments{
\item{am}{association matrix network}

\item{returns}{format for output - either "graph", "matrix" or "edgelist"}

\item{dist.prop}{the proportion of 'rare' ties to reassign. Between zero and one.}
}
\value{

}
\description{
Simulate social distancing using data within days
}
\examples{

}
\author{
Josh Firth
}
