% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{draw.sch.contagion}
\alias{draw.sch.contagion}
\title{Generate school network plot with contagion}
\usage{
draw.sch.contagion(
  am,
  am.layout = NULL,
  layout.type = NULL,
  use.df,
  day,
  groups = NULL
)
}
\arguments{
\item{am}{association matrix}

\item{am.layout}{(if wanting to specify a layout e.g. to keep the layout the same across comparisons)}

\item{use.df}{the dataframe information of the contagion}

\item{day}{the day to do the contagion for}
}
\value{

}
\description{
Generate school network plot with contagion
}
\author{
Josh Firth
}
