% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{format_network}
\alias{format_network}
\title{Matrix to pairwise list without diagonal}
\usage{
format_network(m, idvec = NULL)
}
\arguments{
\item{m}{matrix of contacts}
}
\value{

}
\description{
Matrix to pairwise list without diagonal
}
\examples{

}
\author{
lewis Spurgin
}
