% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for network model}
\usage{
outbreak_setup(net, num.initial.cases, incfn, delayfn, prop.asym, isolation)
}
\arguments{
\item{net}{network in pairwise list format. Object \code{haslemere} is built into the package.}

\item{num.initial.cases}{integer number of initial cases}

\item{incfn}{function that samples from incubation period Weibull distribution; generated using dist_setup}

\item{delayfn}{function that samples from the onset-to-hospitalisation delay Weibull distribution; generated using dist_setup}

\item{prop.asym}{numeric proportion of cases that are sublinical (between 0 and 1)}

\item{isolation}{logical - do you want indiviuals to self-isolate on infection?}
}
\value{
data.frame of cases in outbreak so far
}
\description{
Set up initial cases for network model
}
\examples{

\dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(dist_shape = 1,dist_scale = 1.4)
outbreak_setup(net, num.initial.cases = 5,incfn,delayfn,prop.asym=0, isolation = TRUE)
}


}
\author{
Lewis Spurgin

Joel Hellewell
}
