% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Infections on one day in the network model}
\usage{
outbreak_step(
  day = NULL,
  case_data = NULL,
  net = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  prop.ascertain = NULL,
  presymrate = NULL,
  R = NULL,
  quarantine = NULL,
  isolation = NULL,
  tracing = NULL,
  secondary = NULL,
  outside = NULL,
  sensitivity = NULL,
  testing = NULL,
  cap_max_tests = NULL
)
}
\arguments{
\item{day}{number of days from start of the inection - must be > 0}

\item{case_data}{data.frame of cases in outbreak so far; initially generated by outbreak_setup}

\item{net}{network from which to sample cases, in pairwise list format. Object \code{haslemere} is built into the package}

\item{prop.asym}{proportion of asymptomtaic cases (must be 0<=x<=1)}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{presymrate}{the proportion of cases with pre-onset transmission (must be 0<=x<=1)}

\item{R}{scaling factor for infection probability}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{isolation}{logical whether isolation is in effect}

\item{tracing}{logical whether tracing is in effect}

\item{secondary}{logical whether secondary contact tracing is in effect}

\item{sensitivity}{app sensitivity - "high" or "low"}

\item{testing}{"realistic", "random" or "none"}

\item{cap_max_tests}{integer - maximum number of tests}
}
\value{

}
\description{
Infections on one day in the network model
}
\author{
Lewis Spurgin

Joel Hellewell
}
