% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{plot_network}
\alias{plot_network}
\title{Run an outbreak model and generate network plot with contagion}
\usage{
plot_network(
  am,
  am.layout = NULL,
  day,
  num.initial.cases = NULL,
  prop.ascertain = NULL,
  R = NULL,
  presymrate = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  prop.asym = NULL,
  quarantine = NULL,
  isolation = NULL,
  tracing = NULL,
  secondary = NULL,
  outside = NULL,
  sensitivity = NULL,
  testing = NULL,
  cap_max_tests = NULL,
  s = NULL
)
}
\arguments{
\item{am}{association matrix network}

\item{am.layout}{(if wanting to specify a layout e.g. to keep the layout the same across comparisons)}

\item{day}{the day to do the contagion for}

\item{use.df}{the dataframe information of the contagion}
}
\value{

}
\description{
Run an outbreak model and generate network plot with contagion
}
\examples{

\dontrun{

#example of using function
a<-load(file="/am_list.RData") #need this for a network
b<-load(file="/network_examples.RData") #need this for contagion data


#here, we're going to draw 2 networks, both of the same day (time period of the contagion) and both in the same layout (but  assuming you don't have a pre-specified layout) then just like this works:
dev.new(height=6,width=12);par(mar=c(0,0,0,0));par(mai=c(0,0,0,0));par(mfrow=c(1,2))
am<-am_list[[1]]
#draw your first network and save the layout
lay<-draw.contagion(am,use.df=primary_tracing,day=20)
#now, if you want to use that same layout again for a different network you can just:
draw.contagion(am,am.layout=lay,use.df=secondary_tracing,day=20)
}

}
\author{
Josh Firth

Lewis Spurgin
}
