% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  net = NULL,
  num.initial.cases = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  R = NULL,
  presymrate = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  prop.asym = NULL,
  quarantine = NULL,
  isolation = NULL,
  tracing = NULL,
  secondary = NULL,
  outside = NULL,
  sensitivity = NULL,
  testing = NULL,
  cap_max_tests = NULL,
  weekly = TRUE,
  s = NULL
)
}
\arguments{
\item{net}{network from which to sample cases, in pairwise list format. Object \code{haslemere} is built into the package}

\item{num.initial.cases}{integer number of initial cases}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{R}{scaling factor for infection probability}

\item{presymrate}{the proportion of cases with pre-onset transmission (must be 0<=x<=1)}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{prop.asym}{proportion of asymptomtaic cases (must be 0<=x<=1)}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{isolation}{logical whether isolation is in effect}

\item{tracing}{logical whether tracing is in effect}

\item{secondary}{logical whether secondary contact tracing is in effect}

\item{sensitivity}{app sensitivity - "high" or "low"}

\item{testing}{"realistic", "random" or "none"}

\item{cap_max_tests}{integer - maximum number of tests}

\item{weekly}{logical - return summarised weekly cases or the raw case data}

\item{s}{seed - optional for reproducing the same output. Useful for \code{plot_network()}}
}
\value{
data.frame of cases by week, cumulative cases, and weekly isolations, quarantines and tests
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
outbreak_model(net = haslemere, num.initial.cases = 1, prop.ascertain = 0.8, cap_max_days = 69, R = 6.5, presymrate = 0.4,
delay_shape = 1,delay_scale = 1.4,prop.asym = 0.4, quarantine = TRUE, isolation = TRUE, tracing = TRUE, secondary = TRUE,
outside = 0.001, sensitivity = "high", testing = "none"}


}
\author{
Joel Hellewell

Lewis Spurgin
}
